backend/requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
crewai
